refactor: `NavBar`コンポーネントを分割
